worked on new interrupt system

+++ libc/stdlib/free.c
@@ -25,11 +25,11 @@ void test_free(void)
     int buffer_size = 30;
     char *buffer = malloc(sizeof(char) * buffer_size);
     free(buffer);
-    int old = (uint64_t)buffer;
+    uint64_t old = (uint64_t)buffer;
 
     buffer = malloc(sizeof(char) * buffer_size);
 
-    ctest_assert("Error: malloc,free,malloc produces different pointers", old == buffer);
+    ctest_assert("Error: malloc,free,malloc produces different pointers", old == (uint64_t)buffer);
 
     //not sure what else to test here at the moment
 }
